+++ src/origin.rs
@@ -3,7 +3,6 @@ use num_enum::{TryFromPrimitive, IntoPrimitive};
 use serde_repr::{Serialize_repr, Deserialize_repr};
 
 /// Origin.
-
 #[derive(
     Serialize_repr,
     Deserialize_repr,