+++ src/spell_set.rs
@@ -15,9 +15,7 @@ const SPELL_COUNT: usize = 2;
 /// - An item cannot have duplicate spells or multiple spells of the same type.
 /// - Comparing spells for equality is order-agnostic.
 /// - Hashing is order-agnostic.
-/// 
-/// Since the number of spells is fixed, the struct uses zero heap allocations and is therefore 
-/// [`Copy`].
+/// - The type is `Copy`, allowing for cheap and easy duplication.
 /// 
 /// # Examples
 /// ```
@@ -481,8 +479,9 @@ impl Iterator for SpellSetIterator {
     type Item = Spell;
     
     fn next(&mut self) -> Option<Self::Item> {
-        #[allow(clippy::manual_flatten)]
-        while let Some(opt) = self.inner.next() {
+        let iter = self.inner.by_ref();
+        
+        for opt in iter {
             if opt.is_some() {
                 return opt;
             }