+++ CHANGELOG.md
@@ -11,6 +11,8 @@
 - `from_strange_part_name` method to `StrangePart`.
 - `kill_eater_score_type` method to `StrangePart`.
 - Missing `EnumCount` for some definitions.
+- `defindex_is_strange_part` for `StrangePart`.
+- `defindex_is_paint` for `Paint`.
 - `BitAnd` for `SpellSet` and `StrangePartSet`, and their borrowed variants.
 - `is_subset` and `is_superset` methods to `StrangePartSet` and `SpellSet`.